requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0

# Image Processing
scikit-image>=0.19.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# XAI Comparisons
pytorch-grad-cam>=1.4.0

# Utilities
tqdm>=4.62.0
argparse

# Optional for Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0
